Fix learning drawer accordion latency from context fan-out
Isolate accordion expansion state in a panel-scoped AccordionContext,
pass stable ExpandableSection IDs, and memo leaf consumers so topic
switches no longer grow click latency until E2E stalls.

Resolves: MTV-6166
