+++ src/onlineHelp/learningExperienceDrawer/context/AccordionContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+import type { AccordionContextType } from '../utils/types';
+
+import { createAccordionContext } from './createAccordionContext';
+
+export const AccordionContext = createContext<AccordionContextType>(createAccordionContext());
