+++ src/onlineHelp/learningExperienceDrawer/context/createAccordionContext.ts
@@ -1,0 +1,7 @@
+import type { AccordionContextType } from '../utils/types';
+
+export const createAccordionContext = (): AccordionContextType => ({
+  closeExpansionItem: () => null,
+  openExpansionItem: () => null,
+  openExpansionItems: [],
+});
