+++ src/onlineHelp/learningExperienceDrawer/context/createLearningExperienceContext.ts
@@ -3,13 +3,10 @@ import type { LearningExperienceContextType } from '../utils/types';
 
 export const createLearningExperienceContext = (): LearningExperienceContextType => ({
   clearData: () => null,
-  closeExpansionItem: () => null,
   closeLearningExperience: () => null,
   data: {},
   drawerWidth: DEFAULT_DRAWER_WIDTH,
   isLearningExperienceOpen: false,
-  openExpansionItem: () => null,
-  openExpansionItems: [],
   openLearningExperience: () => null,
   referenceScrollPositions: {},
   scrollPosition: 0,
